+++ src/test/php/ValueContainsTest.php
@@ -1,0 +1,76 @@
+<?php
+/**
+ * This file is part of stubbles.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @package  stubbles\values
+ */
+namespace stubbles\values;
+use function bovigo\assert\assertFalse;
+use function bovigo\assert\assertTrue;
+/**
+ * Tests for stubbles\values\Value::contains().
+ *
+ * @group  values
+ * @group  value_checks
+ * @since  7.2.0
+ */
+class ValueContainsTest extends \PHPUnit_Framework_TestCase
+{
+/**
+     * returns tuples which evaluate to true
+     *
+     * @return  array
+     */
+    public function tuplesEvaluatingToTrue()
+    {
+        return [
+                [null, null],
+                [5, 'foo5'],
+                [5, 'fo5o'],
+                ['foo', 'foobar'],
+                ['foo', 'foo'],
+                ['foo', ['foo', 'bar', 'baz']],
+                [null, ['foo', null, 'baz']]
+        ];
+    }
+
+    /**
+     * @param  mixed                      $needle
+     * @param  string|array|\Traversable  $haystack
+     * @test
+     * @dataProvider  tuplesEvaluatingToTrue
+     */
+    public function evaluatesToTrue($needle, $haystack)
+    {
+        assertTrue(value($haystack)->contains($needle));
+    }
+
+    /**
+     * returns tuples which evaluate to false
+     *
+     * @return  array
+     */
+    public function tuplesEvaluatingToFalse()
+    {
+        return [
+                [5, 'foo'],
+                [true, 'blub'],
+                ['dummy', 'bar'],
+                ['nope', ['foo', 'bar', 'baz']]
+        ];
+    }
+
+    /**
+     * @param  mixed                      $needle
+     * @param  string|array|\Traversable  $haystack
+     * @test
+     * @dataProvider  tuplesEvaluatingToFalse
+     */
+    public function evaluatesToFalse($needle, $haystack)
+    {
+        assertFalse(value($haystack)->contains($needle));
+    }
+}