+++ src/test/php/ValueEqualsTest.php
@@ -1,0 +1,88 @@
+<?php
+/**
+ * This file is part of stubbles.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * @package  stubbles\values
+ */
+namespace stubbles\values;
+use function bovigo\assert\assertFalse;
+use function bovigo\assert\assertTrue;
+use function bovigo\assert\expect;
+/**
+ * Tests for stubbles\values\Value::equals().
+ *
+ * @group  values
+ * @group  value_checks
+ * @since  7.2.0
+ */
+class EqualsTest extends \PHPUnit_Framework_TestCase
+{
+    /**
+     * @test
+     */
+    public function constructionWithObjectThrowsIllegalArgumentException()
+    {
+        $value = value('foo');
+        expect(function() use ($value) {
+                $value->equals(new \stdClass());
+        })->throws(\InvalidArgumentException::class);
+    }
+
+    /**
+     * @return  array
+     */
+    public function tuplesEvaluatingToTrue()
+    {
+        return [[true, true],
+                [false, false],
+                [5, 5],
+                [null, null],
+                ['foo', 'foo']
+        ];
+    }
+
+    /**
+     * @param  scalar  $expected
+     * @param  mixed   $value
+     * @test
+     * @dataProvider  tuplesEvaluatingToTrue
+     */
+    public function evaluatesToTrue($expected, $value)
+    {
+        assertTrue(value($value)->equals($expected));
+    }
+
+    /**
+     * @return  array
+     */
+    public function tuplesEvaluatingToFalse()
+    {
+        return [[true, false],
+                [false, true],
+                [false, new \stdClass()],
+                [false, null],
+                [5, 'foo'],
+                [5, 6],
+                [true, 5],
+                [false, 0],
+                [true, 'foo'],
+                ['foo', 'bar'],
+                [5, new \stdClass()],
+                ['foo', new \stdClass()]
+        ];
+    }
+
+    /**
+     * @param  scalar  $expected
+     * @param  mixed   $value
+     * @test
+     * @dataProvider  tuplesEvaluatingToFalse
+     */
+    public function evaluatesToFalse($expected, $value)
+    {
+        assertFalse(value($value)->equals($expected));
+    }
+}